+++ api/src/routes/auth.route.ts
@@ -4,21 +4,19 @@ import authenticateToken from "../middleware/authenticateToken";
 import { initSchema } from "../validators/query.validator";
 import { validateQuery } from "../validators/validators";
 
-import authenticateTokens from "../middleware/authenticateToken";
-
 const authRoutes = () => {
   const { authorize, refreshAccessToken, login, logout, signup } =
     authController;
   const authRouter = Router();
 
   // GET requests
-  authRouter.get("/", authenticateToken, authorize);
+  authRouter.get("/", authenticateToken(), authorize);
 
   // POST requests
   // curl -w "\n" -X POST http://localhost:3001/auth
   authRouter.post("/login", login);
   authRouter.post("/logout", logout);
-  authRouter.post("/refresh", authenticateTokens, refreshAccessToken);
+  authRouter.post("/refresh", authenticateToken("refresh"), refreshAccessToken);
   authRouter.post("/signup", signup);
 
   // curl -w "\n" -X POST http://localhost:3001/pizza/:id