+++ cms/src/layouts/ProtectedLayout.tsx
@@ -19,23 +19,75 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
 
   useEffect(() => {
     console.log("ProtectedLayout mounted...");
-    //   /* if (!accessToken)
-    //     navigate("/login", { state: { from: location }, replace: true }); */
+    /* if (!accessToken)
+        navigate("/login", { state: { from: location }, replace: true }); */
+    const responseInterceptor = axios.interceptors.response.use(
+      async (response) => {
+        console.log("response in responseInterceptor:", response);
+        /* if (response.data.status === "fail") {
+          return await axios.post("/auth/refresh");
+        } */
+
+        return response;
+      },
+      async (err) => {
+        console.log("err:", err);
+        // return Promise.reject(err);
+        if (err.request.status === 403) {
+          return await axios.post("/auth/refresh");
+        }
+        return Promise.reject(err);
+      }
+    );
+
+    const requestInterceptor = axios.interceptors.request.use((config) => {
+      if (accessToken) {
+        /* axios.defaults.headers.common[
+          "Authorization"
+        ] = `Bearer ${accessToken}`; */
+        config.headers.Authorization = `Bearer ${accessToken}`;
+      }
+      return config;
+    });
+
     const authorize = async () => {
       try {
         console.log("accessToken in authorize():", accessToken);
 
         if (accessToken) {
-          axios.defaults.headers.common[
-            "Authorization"
-          ] = `Bearer ${accessToken}`;
+          console.log("accessToken exists");
+          // axios.defaults.headers.common[
+          //   "Authorization"
+          // ] = `Bearer ${accessToken}`;
+          setTimeout(async () => {
+            console.log('right before await axios.get("/auth") request');
+            console.log("accessToken:", accessToken);
+            // const response = await axios.get("/auth");
+            // console.log("response:", response);
+            // setAccessToken(response.data.data.accessToken);
+          }, 0);
+
+          // const response = await axios.get("/auth");
+          // // Need to intercept response, and send a new request to "/auth/refresh" (post request)
+          // // Only try one attempt each for both paths
+          // // If the response from "/auth/refresh" is bad
+          // //  Then set access token to null and redirect user to login page
+
+          // console.log("response:", response);
+          // setAccessToken(response.data.data.accessToken);
         }
 
-        const response = await axios.post("/auth/refresh");
+        // // https://dikshantraj2001.medium.com/retry-failed-request-using-axios-interceptors-5472549dc57a
+        const response = await axios.get("/auth");
+        // // Need to intercept response, and send a new request to "/auth/refresh" (post request)
+        // // Only try one attempt each for both paths
+        // // If the response from "/auth/refresh" is bad
+        // //  Then set access token to null and redirect user to login page
 
-        console.log("response:", response);
+        console.log("authorize response:", response);
         setAccessToken(response.data.data.accessToken);
       } catch (err) {
+        console.log("authorize err caught");
         console.log(location);
         console.log(err);
         setAccessToken(null);
@@ -44,7 +96,13 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
     };
 
     authorize();
-  }, [location, accessToken]);
+
+    return () => {
+      console.log("ProtectedLayout clean up function running...");
+      axios.interceptors.response.eject(responseInterceptor);
+      axios.interceptors.request.eject(requestInterceptor);
+    };
+  }, [location]);
 
   return (
     <>